make changes search request response

+++ db/migrate/20200116154902_add_avg_rating_to_albums.rb
@@ -1,5 +1,5 @@
 class AddAvgRatingToAlbums < ActiveRecord::Migration[6.0]
   def change
-    add_column :albums, :avg_rating, :integer
+    add_column :albums, :avg_rating, :float
   end
 end
